split cellsnp.c
