+++ Makefile
@@ -10,8 +10,8 @@ BIN_DIR=/usr/local/bin
 BIN_NAME=cellsnp-lite
 
 src_dir=src
-scripts=$(src_dir)/cellsnp.c $(src_dir)/csp.c $(src_dir)/jfile.c $(src_dir)/jsam.c $(src_dir)/jstring.c $(src_dir)/mplp.c $(src_dir)/snp.c $(src_dir)/thpool.c $(src_dir)/thread.c
-headers=$(src_dir)/config.h $(src_dir)/csp.h $(src_dir)/default.h $(src_dir)/jfile.h $(src_dir)/jmemory.h $(src_dir)/jnumeric.h $(src_dir)/jsam.h $(src_dir)/jstring.h $(src_dir)/kvec.h $(src_dir)/mplp.h $(src_dir)/snp.h $(src_dir)/thpool.h $(src_dir)/thread.h
+scripts=$(src_dir)/cellsnp.c $(src_dir)/csp_fetch.c $(src_dir)/csp_pileup.c $(src_dir)/csp.c $(src_dir)/jfile.c $(src_dir)/jsam.c $(src_dir)/jstring.c $(src_dir)/mplp.c $(src_dir)/snp.c $(src_dir)/thpool.c $(src_dir)/thread.c
+headers=$(src_dir)/config.h $(src_dir)/csp.h $(src_dir)/jfile.h $(src_dir)/jmemory.h $(src_dir)/jnumeric.h $(src_dir)/jsam.h $(src_dir)/jstring.h $(src_dir)/kvec.h $(src_dir)/mplp.h $(src_dir)/snp.h $(src_dir)/thpool.h $(src_dir)/thread.h
 
 all: $(BIN_NAME)
